requirements.txt
# requirements.txt (Optimized)
# This file lists the direct dependencies of the project.
# Sub-dependencies are managed by pip during installation.

# --- Core Bot Framework ---
# The main framework for building the Telegram bot.
# The [persistence] extra includes support for saving bot/user data.
python-telegram-bot[persistence]==21.3

# --- Environment & Configuration ---
# Used in config.py to load secrets from a .env file.
python-dotenv==1.1.1

# --- Asynchronous Operations & API Clients ---
# Required for making async HTTP requests to the Ollama API in utils.py.
aiohttp==3.12.13

# Used by the news tool in tools.py to fetch headlines from NewsAPI.
newsapi-python==0.2.7

# Powers the web search tool in tools.py.
duckduckgo-search

# --- Utilities ---
# Used for timezone-aware datetime objects in tools.py (e.g., get_current_time).
pytz